fix: merger gone wrong

+++ internal/controller/httproute_controller.go
@@ -90,8 +90,7 @@ func (r *HTTPRouteReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 				}
 				svcIdx[svc.Name] = svc
 				if ref.Port != nil {
-					portIdx[svc.Name] = int32(*ref.Port)
-)
+					portIdx[svc.Name] = *ref.Port
 				}
 			}
 		}
